# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/groups/{groupId}/people": {
      "get": {
        "operationId": "GroupPeopleController_getPeople",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "includeArchived",
            "required": true,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupPeopleResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List active people represented in a group",
        "tags": [
          "group-people"
        ]
      },
      "post": {
        "operationId": "GroupPeopleController_createPerson",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupPersonBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupPersonDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a placeholder person in a group",
        "tags": [
          "group-people"
        ]
      }
    },
    "/groups/{groupId}/people/catalog": {
      "get": {
        "operationId": "GroupPeopleController_getCatalog",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Optional title search for games that can be assigned to a group person.",
            "schema": {
              "example": "catan",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameCompleteDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List catalog games available for group-person assertions",
        "tags": [
          "group-people"
        ]
      }
    },
    "/groups/{groupId}/people/{personId}": {
      "patch": {
        "operationId": "GroupPeopleController_updatePerson",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "personId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupPersonBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupPersonDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update or archive a group person",
        "tags": [
          "group-people"
        ]
      }
    },
    "/groups/{groupId}/people/{personId}/claim": {
      "post": {
        "operationId": "GroupPeopleController_claimPerson",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "personId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimGroupPersonBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Claim a targeted placeholder as the authenticated account",
        "tags": [
          "group-people"
        ]
      }
    },
    "/groups/{groupId}/people/join": {
      "post": {
        "operationId": "GroupPeopleController_joinAsNewPerson",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupPersonDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Join a group as a new linked group person",
        "tags": [
          "group-people"
        ]
      }
    },
    "/groups/{groupId}/people/{personId}/ownership": {
      "get": {
        "operationId": "GroupPeopleController_getOwnership",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "personId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GroupPersonGameOwnershipDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List group-person game ownership assertions",
        "tags": [
          "group-people"
        ]
      },
      "put": {
        "operationId": "GroupPeopleController_updateOwnership",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "personId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupPersonOwnershipBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ownership assertion saved"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Set a group-person game ownership assertion",
        "tags": [
          "group-people"
        ]
      }
    },
    "/groups/{groupId}/people/{personId}/preferences": {
      "get": {
        "operationId": "GroupPeopleController_getPreferences",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "personId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GroupPersonGamePreferenceDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List group-person game preferences",
        "tags": [
          "group-people"
        ]
      },
      "put": {
        "operationId": "GroupPeopleController_updatePreference",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "personId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupPersonPreferenceBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Preference saved"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Set a group-person game preference",
        "tags": [
          "group-people"
        ]
      }
    },
    "/groups/{groupId}/people/{personId}/preferences/{gameId}": {
      "delete": {
        "operationId": "GroupPeopleController_deletePreference",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "personId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "gameId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Preference removed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a group-person game preference",
        "tags": [
          "group-people"
        ]
      }
    },
...
  },
  "info": {
    "title": "Board Vault API",
    "description": "Versioned contract snapshot for the Board Vault social group workspace API.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AvatarDto": {
        "type": "object",
        "properties": {
          "backgroundColor": {
            "type": "string",
            "example": "#3B82F6"
          },
          "iconName": {
            "type": "object",
            "example": "person-fill"
          },
          "emoji": {
            "type": "object",
            "example": null
          },
          "type": {
            "type": "string",
            "example": "icon"
          },
          "initials": {
            "type": "string",
            "example": "AB"
          }
        },
        "required": [
          "backgroundColor",
          "iconName",
          "emoji",
          "type",
          "initials"
        ]
      },
...
      "GroupPersonDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 42
          },
          "groupId": {
            "type": "number",
            "example": 7
          },
          "accountId": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "kind": {
            "type": "string",
            "enum": [
              "placeholder",
              "linked"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "archived"
            ]
          },
          "displayName": {
            "type": "string",
            "example": "Ana"
          },
          "avatar": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/AvatarDto"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "example": "2026-09-25 12:00:00"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-09-25 12:00:00"
          },
          "claimedAt": {
            "type": "object",
            "example": null,
            "nullable": true
          }
        },
        "required": [
          "id",
          "groupId",
          "accountId",
          "kind",
          "status",
          "displayName",
          "createdAt",
          "updatedAt",
          "claimedAt"
        ]
      },
      "GroupPersonGameOwnershipDto": {
        "type": "object",
        "properties": {
          "gameId": {
            "type": "number",
            "example": 42
          },
          "status": {
            "type": "string",
            "enum": [
              "asserted",
              "rejected",
              "disputed"
            ]
          },
          "source": {
            "type": "string",
            "enum": [
              "placeholder_setup",
              "account_collection",
              "claimed_import"
            ]
          },
          "enteredByAccountId": {
            "type": "number",
            "example": 7
          },
          "confirmedByAccountId": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "example": "2026-09-25 12:00:00"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-09-25 12:00:00"
          }
        },
        "required": [
          "gameId",
          "status",
          "source",
          "enteredByAccountId",
          "confirmedByAccountId",
          "createdAt",
          "updatedAt"
        ]
      },
      "GroupPersonGamePreferenceDto": {
        "type": "object",
        "properties": {
          "gameId": {
            "type": "number",
            "example": 42
          },
          "preference": {
            "type": "string",
            "enum": [
              "favorite",
              "like",
              "neutral",
              "avoid"
            ]
          },
          "source": {
            "type": "string",
            "enum": [
              "placeholder_setup",
              "claimed_import",
              "account_profile"
            ]
          },
          "enteredByAccountId": {
            "type": "number",
            "example": 7
          },
          "createdAt": {
            "type": "string",
            "example": "2026-09-25 12:00:00"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-09-25 12:00:00"
          }
        },
        "required": [
          "gameId",
          "preference",
          "source",
          "enteredByAccountId",
          "createdAt",
          "updatedAt"
        ]
      },
      "GroupPersonWorkspaceDto": {
        "type": "object",
        "properties": {
          "person": {
            "$ref": "#/components/schemas/GroupPersonDto"
          },
          "ownership": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupPersonGameOwnershipDto"
            }
          },
          "preferences": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupPersonGamePreferenceDto"
            }
          },
          "claimable": {
            "type": "boolean",
            "example": false,
            "description": "Whether the authenticated account may claim this placeholder."
          }
        },
        "required": [
          "person",
          "ownership",
          "preferences",
          "claimable"
        ]
      },
      "GroupPeopleResponseDto": {
        "type": "object",
        "properties": {
          "people": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupPersonWorkspaceDto"
            }
          }
        },
        "required": [
          "people"
        ]
      },
      "GameCompleteDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 12345
          },
          "imageUrl": {
            "type": "string",
            "example": "https://www.example.com/image.jpg"
          },
          "gameAvgDuration": {
            "type": "number",
            "example": 120,
            "description": "The average duration of the game in minutes."
          },
          "minPlayers": {
            "type": "number",
            "example": 3,
            "description": "The minimum number of players required to play the game."
          },
          "maxPlayers": {
            "type": "number",
            "example": 4,
            "description": "The maximum number of players that can play the game."
          },
          "title": {
            "type": "string",
            "example": "Game Title"
          },
          "titleTranslations": {
            "type": "object",
            "example": {
              "en": "Game Title"
            }
          }
        },
        "required": [
          "id",
          "imageUrl",
          "gameAvgDuration",
          "minPlayers",
          "maxPlayers",
          "titleTranslations"
        ]
      },
      "CreateGroupPersonBody": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "example": "Ana"
          },
          "avatar": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/AvatarDto"
              }
            ]
          }
        },
        "required": [
          "displayName"
        ]
      },
      "UpdateGroupPersonBody": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "example": "Ana"
          },
          "avatar": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/AvatarDto"
              }
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "archived"
            ]
          }
        }
      },
      "ClaimGroupPersonBody": {
        "type": "object",
        "properties": {
          "ownershipGameIds": {
            "example": [
              42,
              84
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "preferenceGameIds": {
            "example": [
              42
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "importOwnershipToCollection": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "UpdateGroupPersonOwnershipBody": {
        "type": "object",
        "properties": {
          "gameId": {
            "type": "number",
            "example": 42
          },
          "status": {
            "type": "string",
            "enum": [
              "asserted",
              "rejected",
              "disputed"
            ]
          }
        },
        "required": [
          "gameId",
          "status"
        ]
      },
      "UpdateGroupPersonPreferenceBody": {
        "type": "object",
        "properties": {
          "gameId": {
            "type": "number",
            "example": 42
          },
          "preference": {
            "type": "string",
            "enum": [
              "favorite",
              "like",
              "neutral",
              "avoid"
            ]
          }
        },
        "required": [
          "gameId",
          "preference"
        ]
      },
...
    }
  }
}